+++ src/configurationProvider.ts
@@ -1,8 +1,5 @@
 const ConfigParser = require("configparser");
-
-// import * as path from "path";
 const path = require("path");
-import * as fs from "fs";
 
 export class RepositoryInformationManager {
     public token(rootPath: string | undefined) {
@@ -16,15 +13,12 @@ export class RepositoryInformationManager {
         const domain = config.get("REPO", "domain");
         const repo_owner = config.get("REPO", "repo_owner");
         const repo_name = config.get("REPO", "repo_name");
-        return "http://" + domain + "/api/v1/repos/" + repo_owner + "/" + repo_name + "/issues";
+        const ssl = config.get("REPO", "ssl");
+        const prefix = ssl ? "https" : "http";
+        return prefix + "://" + domain + "/api/v1/repos/" + repo_owner + "/" + repo_name + "/issues";
     }
     public saveRepoInformation(rootPath: string | undefined, repoInformations: any) {
         const file_path = path.join(rootPath as string, ".gitea/config.ini");
-        try {
-            fs.mkdirSync(path.join(rootPath as string, ".gitea/"));
-        } catch (error) {
-
-        }
         const config = new ConfigParser();
         config.addSection("PRIVATE");
         config.addSection("REPO");
@@ -32,6 +26,7 @@ export class RepositoryInformationManager {
         config.set("REPO", "domain", repoInformations.domain);
         config.set("REPO", "repo_owner", repoInformations.repo_owner);
         config.set("REPO", "repo_name", repoInformations.repo_name);
-        config.write(file_path);
+        config.set("REPO", "ssl", true);
+        config.write(file_path, true);
     }
 }